testes automaticos

+++ workspace/desafioCin/src/main/java/br/com/desafio/cin/samsung/controles/EquipamentoController.java
@@ -1,6 +1,10 @@
 package br.com.desafio.cin.samsung.controles;
 
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.Optional;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -30,43 +34,46 @@ import br.com.desafio.cin.samsung.utils.QrCode;
 @RequestMapping("/api/equipamento")
 @CrossOrigin(origins = "*")
 public class EquipamentoController {
-	
+
 	@Autowired
 	private EquipamentoService equipamentoService;
-	
-	@PostMapping
-	public ResponseEntity<Response<Equipamento>> create(HttpServletRequest request, @RequestBody Equipamento equipamento,
-			BindingResult result) {
-		Response<Equipamento> Response = new Response<Equipamento>();
-		
+
+	@PostMapping(value = "criar")
+	public ResponseEntity<Response<Equipamento>> create(HttpServletRequest request,
+			@RequestBody Equipamento equipamento, BindingResult result) {
+		Response<Equipamento> response = new Response<Equipamento>();
+
 		QrCode.removerImagensDiretorio();
-		
+
 		try {
 			validateComun(equipamento, result);
 			if (result.hasErrors()) {
-				result.getAllErrors().forEach(erro -> Response.getErros().add(erro.getDefaultMessage()));
-				return ResponseEntity.badRequest().body(Response);
+				result.getAllErrors().forEach(erro -> response.getErros().add(erro.getDefaultMessage()));
+				return ResponseEntity.badRequest().body(response);
 			}
-			
+
 			gerarQrCode(equipamento);
-		
+			
 			Equipamento newEquipamento = equipamentoService.createOrUpdate(equipamento);
-			Response.setData(newEquipamento);
+			response.setData(newEquipamento);
 
 		} catch (Exception e) {
-			Response.getErros().add(e.getMessage());
+			response.getErros().add(e.getMessage());
 		}
-		return ResponseEntity.ok(Response);
+		return ResponseEntity.ok(response);
 	}
 
-	private void gerarQrCode(Equipamento equipamento) {
-		ObjectMapper Obj = new ObjectMapper(); 
+	private FileOutputStream gerarQrCode(Equipamento equipamento) {
+		ObjectMapper Obj = new ObjectMapper();
 		int tamanho = 125;
+		FileOutputStream qrcode = null;
 		try {
-			QrCode.criarQRCode(Obj.writeValueAsString(equipamento), tamanho);
+			qrcode = QrCode.criarQRCode(Obj.writeValueAsString(equipamento), tamanho);
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
+		
+		return qrcode;
 	}
 
 	private void validateComun(Equipamento equipamento, BindingResult result) {
@@ -74,7 +81,7 @@ public class EquipamentoController {
 		if (equipamento.getTipo() == null) {
 			result.addError(new ObjectError(Equipamento.class.getSimpleName(), "Tipo é obrigatório."));
 		}
-		
+
 		if (equipamento.getModelo() == null || "".equals(equipamento.getModelo())) {
 			result.addError(new ObjectError(Equipamento.class.getSimpleName(), "Modelo é obrigatório."));
 		}
@@ -82,23 +89,24 @@ public class EquipamentoController {
 		if (equipamento.getMesano() == null) {
 			result.addError(new ObjectError(Equipamento.class.getSimpleName(), "Mês ano é obrigatório."));
 		}
-		
+
 		if (equipamento.getValor() == null || equipamento.getValor().equals(Double.MIN_VALUE)) {
 			result.addError(new ObjectError(Equipamento.class.getSimpleName(), "Valor é obrigatório."));
 		}
-		
+
 		if (equipamento.getFoto() == null) {
 			result.addError(new ObjectError(Equipamento.class.getSimpleName(), "Foto é obrigatório."));
 		}
+
 	}
 
-	@PutMapping
-	public ResponseEntity<Response<Equipamento>> update(HttpServletRequest request, @RequestBody Equipamento equipamento,
-			BindingResult result) {
+	@PutMapping(value = "atualizar")
+	public ResponseEntity<Response<Equipamento>> update(HttpServletRequest request,
+			@RequestBody Equipamento equipamento, BindingResult result) {
 		Response<Equipamento> Response = new Response<Equipamento>();
 
 		QrCode.removerImagensDiretorio();
-		
+
 		try {
 			validateUpdate(equipamento, result);
 			validateComun(equipamento, result);
@@ -106,62 +114,75 @@ public class EquipamentoController {
 				result.getAllErrors().forEach(erro -> Response.getErros().add(erro.getDefaultMessage()));
 				return ResponseEntity.badRequest().body(Response);
 			}
-			
+
 			gerarQrCode(equipamento);
-			
+
 			Equipamento equipamentoUpdated = equipamentoService.createOrUpdate(equipamento);
 			Response.setData(equipamentoUpdated);
 
 		} catch (Exception e) {
 			Response.getErros().add(e.getMessage());
-			ResponseEntity.badRequest().body(Response);
+			return ResponseEntity.badRequest().body(Response);
 		}
 		return ResponseEntity.ok(Response);
 	}
 
 	private void validateUpdate(Equipamento Equipamento, BindingResult result) {
-		
+
 		if (Equipamento.getIdEquipamento() == null) {
 			result.addError(new ObjectError(Equipamento.class.getSimpleName(), "Id Equipamento é obrigatório."));
-		}		
+		}
 	}
-	
+
 	@GetMapping(value = "{id}")
 	public ResponseEntity<Response<Equipamento>> findById(@PathVariable("id") String idEquipamento) {
 		Response<Equipamento> response = new Response<Equipamento>();
 
-		Equipamento equipamento = equipamentoService.findByIdEquipamento(new Long(idEquipamento));
+		Optional<Equipamento> equipamento = equipamentoService.findByIdEquipamento(new Long(idEquipamento));
 
-		if (equipamento == null) {
+		if (equipamento == null || !equipamento.isPresent()) {
 			response.getErros().add("Equipamento não encontrado. IdEquipamento = " + idEquipamento);
-			ResponseEntity.badRequest().body(response);
+			return ResponseEntity.badRequest().body(response);
 		}
-		response.setData(equipamento);
+		response.setData(equipamento.get());
 		return ResponseEntity.ok().body(response);
 	}
-	
+
 	@DeleteMapping(value = "{id}")
 	public ResponseEntity<Response<String>> delete(@PathVariable("id") String idEquipamento) {
 		Response<String> response = new Response<String>();
+		
+		if (idEquipamento == null || "".equalsIgnoreCase(idEquipamento) || "null".equalsIgnoreCase(idEquipamento)) {
+			response.getErros().add("IdEquipamento = " + idEquipamento + " não pode ser nulo.");
+			return ResponseEntity.badRequest().body(response);
+		}
 
-		Equipamento Equipamento = equipamentoService.findByIdEquipamento(new Long(idEquipamento));
-		if (Equipamento == null) {
-			response.getErros().add("Equipamento not found. IdEquipamento = " + idEquipamento);
-			ResponseEntity.badRequest().body(response);
+		Optional<Equipamento> equipamento = equipamentoService.findByIdEquipamento(new Long(idEquipamento));
+
+		if (equipamento == null || !equipamento.isPresent()) {
+			response.getErros().add("Equipamento não encontrado. IdEquipamento = " + idEquipamento);
+			return ResponseEntity.badRequest().body(response);
 		}
 
 		equipamentoService.delete(new Long(idEquipamento));
 
 		return ResponseEntity.ok(new Response<String>());
 	}
-	
+
 	@GetMapping(value = "{page}/{count}")
 	public ResponseEntity<Response<Page<Equipamento>>> findAll(@PathVariable("page") int page,
 			@PathVariable("count") int count) {
 		Response<Page<Equipamento>> response = new Response<Page<Equipamento>>();
+
+		if (page == 0 && count == 0) {
+			response.getErros().add("Falha na paginação");
+			return ResponseEntity.badRequest().body(response);
+		}
+
 		Page<Equipamento> equipamento = equipamentoService.findAll(page, count);
 		response.setData(equipamento);
 		return ResponseEntity.ok().body(response);
+
 	}
-	
+
 }
